requirements.txt
aiogram>=3.0,<4.0
python-dotenv>=1.0.0